requirements.txt
asgiref==3.7.2
Django==4.2.1
django-bootstrap-datepicker-plus==5.0.3
django-tinymce==3.6.1
EasyProcess==1.1
entrypoint2==1.1
guardian==0.2.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
Pillow==9.5.0
PyAutoGUI==0.9.54
pydantic==1.10.9
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-http-client==3.3.7
pytweening==1.0.7
PyYAML==6.0
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.6.3
tzdata==2023.3
